+++ components-starters/components-starter-port-shift/src/main/java/com/zoe/framework/components/starter/portShift/configuration/PortShiftAutoConfiguration.java
@@ -42,7 +42,7 @@ public class PortShiftAutoConfiguration {
             ServerProperties serverProperties,
             GlobalPortProperties globalPortProperties) {
         return factory -> {
-            if (!globalPortProperties.getEnable()) {
+            if (!globalPortProperties.isEnable()) {
                 return;
             }
             int basePort = serverProperties.getPort();